+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file.
 * Inline HTML fetching/normalization into BaseMarkdownFetcher and remove fetch/html.py
 * Rename Markdown fetcher modules/classes to HTML fetcher naming and update imports
 * Move test progress runner from tests/support into crawl4md.utils and update imports
+* Move preprocessing RuleCase/data-provider helpers into crawl4md.utils and update test imports
 
 ### Test